+++ stenwiki/make.bat
